pyproject.toml
[tool.poetry]
name = "fastapi project for s&p 500 stocks price prediction"
version = "0.1.0"
description = ""
authors = ["Alexander <alexandr.tereshin@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.29"
yfinance = "^0.2.37"
pandas = "^2.2.1"
beautifulsoup4 = "^4.12.3"
ta = "^0.11.0"
catboost = "^1.2.3"
scikit-learn = "^1.4.1.post1"
category-encoders = "^2.6.3"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
